generator-lab/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
boto3
pypdf
pydantic-settings
